refactor: LadderRung 캐시 도입 및 패키지 구조 변경
